[ide-service] cache ide manifest
